Address comments on SD trace exporter unit test.
